refactor(broker): extract run_interactive_session to eliminate daemon duplication
Move the interactive client handling logic (host election, history
replay, join/leave events, inbound/outbound message loops, cleanup)
from handle_client into a shared pub(crate) run_interactive_session
function. Both the single-room broker and daemon dispatcher now call
this shared function instead of duplicating ~200 lines.

Net change: -193 lines.
