+++ crates/room-cli/src/broker/daemon.rs
@@ -30,7 +30,6 @@ use std::{
 };
 
 use tokio::{
-    io::BufReader,
     net::UnixListener,
     sync::{broadcast, watch, Mutex},
 };
@@ -322,229 +321,11 @@ async fn dispatch_connection(
         .await
         .insert(cid, (String::new(), tx.clone()));
 
-    // Re-assemble the stream for handle_client. We already consumed the first
-    // line, so we inject the username by pre-registering it in the client map
-    // and calling handle_client_interactive directly.
-    let result = handle_client_after_handshake(cid, username, reader, write_half, tx, &state).await;
+    let result =
+        super::run_interactive_session(cid, username, reader, write_half, tx, &state).await;
 
     state.clients.lock().await.remove(&cid);
     result
-}
-
-/// Handle an interactive client after the handshake line has already been
-/// consumed and parsed. This is the same logic as `handle_client` in mod.rs
-/// but skips reading the first line since the daemon dispatcher already did.
-async fn handle_client_after_handshake(
-    cid: u64,
-    username: &str,
-    reader: BufReader<tokio::net::unix::OwnedReadHalf>,
-    write_half: tokio::net::unix::OwnedWriteHalf,
-    own_tx: broadcast::Sender<String>,
-    state: &Arc<RoomState>,
-) -> anyhow::Result<()> {
-    use tokio::io::AsyncWriteExt;
-
-    let username = username.to_owned();
-
-    // Register username in the client map.
-    {
-        let mut map = state.clients.lock().await;
-        if let Some(entry) = map.get_mut(&cid) {
-            entry.0 = username.clone();
-        }
-    }
-
-    // Register as host if no host has been set yet.
-    {
-        let mut host = state.host_user.lock().await;
-        if host.is_none() {
-            *host = Some(username.clone());
-        }
-    }
-
-    eprintln!(
-        "[daemon] {username} joined room {} (cid={cid})",
-        state.room_id
-    );
-
-    // Track this user in the status map.
-    state
-        .status_map
-        .lock()
-        .await
-        .insert(username.clone(), String::new());
-
-    // Subscribe before sending history.
-    let mut rx = own_tx.subscribe();
-
-    // Send chat history, filtering DMs.
-    let host_name = state.host_user.lock().await.clone();
-    let is_host = host_name.as_deref() == Some(username.as_str());
-    let history = crate::history::load(&state.chat_path)
-        .await
-        .unwrap_or_default();
-    let mut write_half = write_half;
-    for msg in &history {
-        let visible = match msg {
-            crate::message::Message::DirectMessage { user, to, .. } => {
-                is_host || user == &username || to == &username
-            }
-            _ => true,
-        };
-        if visible {
-            let line = format!("{}\n", serde_json::to_string(msg)?);
-            if write_half.write_all(line.as_bytes()).await.is_err() {
-                return Ok(());
-            }
-        }
-    }
-
-    // Broadcast join event.
-    let join_msg = crate::message::make_join(&state.room_id, &username);
-    if let Err(e) = super::fanout::broadcast_and_persist(
-        &join_msg,
-        &state.clients,
-        &state.chat_path,
-        &state.seq_counter,
-    )
-    .await
-    {
-        eprintln!("[daemon] broadcast_and_persist(join) failed: {e:#}");
-        return Ok(());
-    }
-
-    // Wrap write half for shared use.
-    let write_half = Arc::new(Mutex::new(write_half));
-
-    // Outbound task.
-    let write_half_out = write_half.clone();
-    let mut shutdown_rx = state.shutdown.subscribe();
-    let outbound = tokio::spawn(async move {
-        loop {
-            tokio::select! {
-                result = rx.recv() => {
-                    match result {
-                        Ok(line) => {
-                            let mut wh = write_half_out.lock().await;
-                            if wh.write_all(line.as_bytes()).await.is_err() {
-                                break;
-                            }
-                        }
-                        Err(broadcast::error::RecvError::Lagged(n)) => {
-                            eprintln!("[daemon] cid={cid} lagged by {n}");
-                        }
-                        Err(broadcast::error::RecvError::Closed) => break,
-                    }
-                }
-                _ = shutdown_rx.changed() => {
-                    while let Ok(line) = rx.try_recv() {
-                        let mut wh = write_half_out.lock().await;
-                        let _ = wh.write_all(line.as_bytes()).await;
-                    }
-                    let _ = write_half_out.lock().await.shutdown().await;
-                    break;
-                }
-            }
-        }
-    });
-
-    // Inbound task.
-    let username_in = username.clone();
-    let room_id_in = state.room_id.clone();
-    let write_half_in = write_half.clone();
-    let state_in = state.clone();
-    let inbound = tokio::spawn(async move {
-        use tokio::io::AsyncBufReadExt;
-
-        let mut reader = reader;
-        let mut line = String::new();
-        loop {
-            line.clear();
-            match reader.read_line(&mut line).await {
-                Ok(0) => break,
-                Ok(_) => {
-                    let trimmed = line.trim();
-                    if trimmed.is_empty() {
-                        continue;
-                    }
-                    match crate::message::parse_client_line(trimmed, &room_id_in, &username_in) {
-                        Ok(msg) => {
-                            match super::commands::route_command(msg, &username_in, &state_in).await
-                            {
-                                Ok(
-                                    super::commands::CommandResult::Handled
-                                    | super::commands::CommandResult::HandledWithReply(_),
-                                ) => {}
-                                Ok(super::commands::CommandResult::Reply(json)) => {
-                                    let _ = write_half_in
-                                        .lock()
-                                        .await
-                                        .write_all(format!("{json}\n").as_bytes())
-                                        .await;
-                                }
-                                Ok(super::commands::CommandResult::Shutdown) => break,
-                                Ok(super::commands::CommandResult::Passthrough(msg)) => {
-                                    let result = match &msg {
-                                        crate::message::Message::DirectMessage { to, .. } => {
-                                            super::fanout::dm_and_persist(
-                                                &msg,
-                                                &username_in,
-                                                to,
-                                                &state_in.host_user,
-                                                &state_in.clients,
-                                                &state_in.chat_path,
-                                                &state_in.seq_counter,
-                                            )
-                                            .await
-                                        }
-                                        _ => {
-                                            super::fanout::broadcast_and_persist(
-                                                &msg,
-                                                &state_in.clients,
-                                                &state_in.chat_path,
-                                                &state_in.seq_counter,
-                                            )
-                                            .await
-                                        }
-                                    };
-                                    if let Err(e) = result {
-                                        eprintln!("[daemon] persist error: {e:#}");
-                                    }
-                                }
-                                Err(e) => eprintln!("[daemon] route error: {e:#}"),
-                            }
-                        }
-                        Err(e) => {
-                            eprintln!("[daemon] bad message from {username_in}: {e}");
-                        }
-                    }
-                }
-                Err(_) => break,
-            }
-        }
-    });
-
-    tokio::select! {
-        _ = outbound => {},
-        _ = inbound => {},
-    }
-
-    // Cleanup.
-    state.status_map.lock().await.remove(&username);
-    let leave_msg = crate::message::make_leave(&state.room_id, &username);
-    let _ = super::fanout::broadcast_and_persist(
-        &leave_msg,
-        &state.clients,
-        &state.chat_path,
-        &state.seq_counter,
-    )
-    .await;
-    eprintln!(
-        "[daemon] {username} left room {} (cid={cid})",
-        state.room_id
-    );
-
-    Ok(())
 }
 
 // ── Tests ─────────────────────────────────────────────────────────────────────
